Added Velocity feature to PlanetDragController
