+++ Assets/Materials/Lighting_ProceduralCelShader.mat
@@ -7,8 +7,8 @@ Material:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_Name: RampDIffuseCelShader
-  m_Shader: {fileID: 4800000, guid: 738fd4cc748cf417098c0a333cbee743, type: 3}
+  m_Name: Lighting_ProceduralCelShader
+  m_Shader: {fileID: 4800000, guid: f811cf26c51b742fbb5b47c12b23e587, type: 3}
   m_ShaderKeywords: 
   m_LightmapFlags: 4
   m_EnableInstancingVariants: 0
@@ -18,11 +18,8 @@ Material:
   disabledShaderPasses: []
   m_SavedProperties:
     serializedVersion: 3
-    m_TexEnvs:
-    - _RampTex:
-        m_Texture: {fileID: 2800000, guid: 3b487d7e90f7f4ab4af2d6ec518591d6, type: 3}
-        m_Scale: {x: 1, y: 1}
-        m_Offset: {x: 0, y: 0}
-    m_Floats: []
+    m_TexEnvs: []
+    m_Floats:
+    - _nDiv: 9.5
     m_Colors:
-    - _Color: {r: 0.6226415, g: 0.44348523, b: 0.57981884, a: 1}
+    - _Color: {r: 0.26210395, g: 0.44131193, b: 0.5849056, a: 1}
