+++ Assets/Materials/Lighting_ProceduralCelShader.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1041d44e6b6e24552a32f7baf27da1b5
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
