+++ Assets/Materials/Lighting_RampNormalCelShader.mat
@@ -7,8 +7,8 @@ Material:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_Name: RampNormalCelShader
-  m_Shader: {fileID: 4800000, guid: 3e8da148de7fc45368ad76ebecdf9f91, type: 3}
+  m_Name: Lighting_RampNormalCelShader
+  m_Shader: {fileID: 4800000, guid: 07e3efe476ee04bdda5277e814505260, type: 3}
   m_ShaderKeywords: 
   m_LightmapFlags: 4
   m_EnableInstancingVariants: 0
@@ -20,18 +20,18 @@ Material:
     serializedVersion: 3
     m_TexEnvs:
     - _BumpMap:
-        m_Texture: {fileID: 2800000, guid: fcc28399a502c4499b5b2e789fc92b36, type: 3}
+        m_Texture: {fileID: 2800000, guid: 95714263336d24633985367c76aa267d, type: 3}
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0, y: 0}
     - _DiffuseTex:
-        m_Texture: {fileID: 2800000, guid: acfc72e9ee06e4d538b45d4a1d51c69b, type: 3}
+        m_Texture: {fileID: 2800000, guid: 1853a70f3dc7c4b449155894e0bbc510, type: 3}
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0, y: 0}
     - _RampTex:
-        m_Texture: {fileID: 2800000, guid: f61835ce2eec14467a40121820b6e494, type: 3}
+        m_Texture: {fileID: 2800000, guid: 3bce270cb303c4f6ebec48674aaa7b29, type: 3}
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0, y: 0}
     m_Floats:
-    - _BumpMultiplier: 1.1
+    - _BumpMultiplier: 0.8
     m_Colors:
     - _Color: {r: 1, g: 1, b: 1, a: 1}
