+++ Assets/Materials/Lighting_RampNormalCelShader.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a8c7dc44a42ab45e383384313b037b14
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
