+++ Assets/Materials/RampDiffuseCelShader.shader.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: f7f2a4bbbfe46448db429e65471161b8
+ShaderImporter:
+  externalObjects: {}
+  defaultTextures: []
+  nonModifiableTextures: []
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
