+++ Assets/Materials/RampNormalCelShader.shader.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 07e3efe476ee04bdda5277e814505260
+ShaderImporter:
+  externalObjects: {}
+  defaultTextures: []
+  nonModifiableTextures: []
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
