+++ Assets/Materials/WindWakeCelShader.shader.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: b5f0b7448ff5f4005b743f2f8315bc21
+ShaderImporter:
+  externalObjects: {}
+  defaultTextures: []
+  nonModifiableTextures: []
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
