+++ Assets/Materials/proceduralCelShader.shader.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: f811cf26c51b742fbb5b47c12b23e587
+ShaderImporter:
+  externalObjects: {}
+  defaultTextures: []
+  nonModifiableTextures: []
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
