+++ Assets/Models/AntAnimator.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b41ebdff4ca0a40a7810af5bf0af2025
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
