+++ Assets/Models/antRigOne.blend.meta
@@ -1,0 +1,149 @@
+fileFormatVersion: 2
+guid: 32c17a40ad03b41ba9d71f488c600adf
+ModelImporter:
+  serializedVersion: 23
+  fileIDToRecycleName:
+    100000: AntModel
+    100002: //RootNode
+    100004: Armature
+    100006: BackLower_L
+    100008: BackLower_R
+    100010: BackMiddle_L
+    100012: BackMiddle_R
+    100014: BackUpper_L
+    100016: BackUpper_R
+    100018: Bone
+    100020: FrontLower_L
+    100022: FrontLower_R
+    100024: FrontMiddle_L
+    100026: FrontMiddle_R
+    100028: FrontUpper_L
+    100030: FrontUpper_R
+    100032: Head
+    100034: MiddleLower_L
+    100036: MiddleLower_R
+    100038: MiddleMiddle_L
+    100040: MiddleMiddle_R
+    100042: MiddleUpper_L
+    100044: MiddleUpper_R
+    100046: Neckish
+    100048: Root
+    100050: Thorax
+    400000: AntModel
+    400002: //RootNode
+    400004: Armature
+    400006: BackLower_L
+    400008: BackLower_R
+    400010: BackMiddle_L
+    400012: BackMiddle_R
+    400014: BackUpper_L
+    400016: BackUpper_R
+    400018: Bone
+    400020: FrontLower_L
+    400022: FrontLower_R
+    400024: FrontMiddle_L
+    400026: FrontMiddle_R
+    400028: FrontUpper_L
+    400030: FrontUpper_R
+    400032: Head
+    400034: MiddleLower_L
+    400036: MiddleLower_R
+    400038: MiddleMiddle_L
+    400040: MiddleMiddle_R
+    400042: MiddleUpper_L
+    400044: MiddleUpper_R
+    400046: Neckish
+    400048: Root
+    400050: Thorax
+    2100000: Material
+    4300000: AntModel
+    7400000: Default Take
+    7400002: walkwalk
+    9500000: //RootNode
+    13700000: AntModel
+  externalObjects: {}
+  materials:
+    importMaterials: 1
+    materialName: 0
+    materialSearch: 1
+    materialLocation: 1
+  animations:
+    legacyGenerateAnimations: 4
+    bakeSimulation: 0
+    resampleCurves: 1
+    optimizeGameObjects: 0
+    motionNodeName: 
+    rigImportErrors: 
+    rigImportWarnings: 
+    animationImportErrors: 
+    animationImportWarnings: 
+    animationRetargetingWarnings: 
+    animationDoRetargetingWarnings: 0
+    importAnimatedCustomProperties: 0
+    importConstraints: 0
+    animationCompression: 1
+    animationRotationError: 0.5
+    animationPositionError: 0.5
+    animationScaleError: 0.5
+    animationWrapMode: 0
+    extraExposedTransformPaths: []
+    extraUserProperties: []
+    clipAnimations: []
+    isReadable: 1
+  meshes:
+    lODScreenPercentages: []
+    globalScale: 1
+    meshCompression: 0
+    addColliders: 0
+    useSRGBMaterialColor: 1
+    importVisibility: 1
+    importBlendShapes: 1
+    importCameras: 1
+    importLights: 1
+    swapUVChannels: 0
+    generateSecondaryUV: 0
+    useFileUnits: 1
+    optimizeMeshForGPU: 1
+    keepQuads: 0
+    weldVertices: 1
+    preserveHierarchy: 0
+    indexFormat: 0
+    secondaryUVAngleDistortion: 8
+    secondaryUVAreaDistortion: 15.000001
+    secondaryUVHardAngle: 88
+    secondaryUVPackMargin: 4
+    useFileScale: 1
+    previousCalculatedGlobalScale: 1
+    hasPreviousCalculatedGlobalScale: 1
+  tangentSpace:
+    normalSmoothAngle: 60
+    normalImportMode: 0
+    tangentImportMode: 3
+    normalCalculationMode: 4
+    legacyComputeAllNormalsFromSmoothingGroupsWhenMeshHasBlendShapes: 0
+    blendShapeNormalImportMode: 1
+    normalSmoothingSource: 0
+  importAnimation: 1
+  copyAvatar: 0
+  humanDescription:
+    serializedVersion: 2
+    human: []
+    skeleton: []
+    armTwist: 0.5
+    foreArmTwist: 0.5
+    upperLegTwist: 0.5
+    legTwist: 0.5
+    armStretch: 0.05
+    legStretch: 0.05
+    feetSpacing: 0
+    rootMotionBoneName: 
+    hasTranslationDoF: 0
+    hasExtraRoot: 0
+    skeletonHasParents: 1
+  lastHumanDescriptionAvatarSource: {instanceID: 0}
+  animationType: 2
+  humanoidOversampling: 1
+  additionalBone: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
