+++ Assets/Scenes/SpinTest.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 35ced74abc7aa411ca91460f24a1c7f1
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
