+++ Assets/Textures.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 48607bebb6936452eaf4dd912e16329c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
