--- UnityPackageManager/manifest.json
@@ -1,4 +1,0 @@
-{
-	"dependencies": {
-	}
-}
